Modified rows and columns
Row and Columns titles now use alphabet letters to donate position as this now allows for 26x26 grids fitting the current cell size of 1 character. Additionally there was an error in how processing the coordinates of the chosen cell worked meaning it could only handle inputs less than 10, this is now fixed. This does however need to be updated to handle letter inputs