Player Winning
Added method to determine if player has won the game, triggers when all mines have been flagged, then checks to see if the player flagged an incorrect mine. The corresponding end message is displayed. Additionally there are some slight fixes to flagging and the order in which some messages take place